// src/app/models/tutor.ts
export interface Tutor {
    IdTutor?: number;
    IdTutorRol?: number;
    IdAlumno?: number;
    NombreTutor?: string;
    PaternoTutor?: string;
    MaternoTutor?: string;
    TutorLegal?: boolean;
    IdEstadoTutor?: number;
    CURPTutor?: string;
    INETutor?: string;
    SexoTutor?: string;
    NacimientoTutor?: Date;
    CalleTutor?: string;
    NumeroTutor?: string;
    IdColoniaTutor?: number;
    MunicipioTutor?: string;
    PostalTutor?: number;
    StatusTutor?: boolean;
    FotoTutor?: string;
    TelefonoTutor?: number;
    CorreoTutor?: string;
    CreatedAt?: Date;
}