// src/app/models/papa.ts
export interface Papa {
    IdPapa?: number;
    IdAlumno?: number;
    PapaNombre?: string;
    PapaPaterno?: string;
    PapaMaterno?: string;
    PapaTutorLegal?: Boolean;
    PapaIdEstado?: number;
    PapaCURP?: string;
    PapaINE?: string;
    PapaSexo?: string;
    PapaNacimiento?: Date;
    PapaCalle?: string;
	PapaNumero?: string;
	PapaIdColonia?: number;
	PapaMunicipio?: string;
	PapaPostal?: number;
    PapaStatus?: boolean;
    PapaFoto?: string;
    PapaTelefono?: number;
    PapaCorreo?: string;
    PapaCreatedAt?: Date;
}